+++ tests/system/ext_test.php
@@ -28,18 +28,13 @@ class ext_test extends \phpbb_test_case
 		parent::setUp();
 
 		// Stub the container
-		$this->container = $this->getMockBuilder('\Symfony\Component\DependencyInjection\ContainerInterface')
-			->getMock();
+		$this->container = $this->createMock('\Symfony\Component\DependencyInjection\ContainerInterface');
 
 		// Stub the ext finder and disable its constructor
-		$this->extension_finder = $this->getMockBuilder('\phpbb\finder')
-			->disableOriginalConstructor()
-			->getMock();
+		$this->extension_finder = $this->createMock('\phpbb\finder');
 
 		// Stub the migrator and disable its constructor
-		$this->migrator = $this->getMockBuilder('\phpbb\db\migrator')
-			->disableOriginalConstructor()
-			->getMock();
+		$this->migrator = $this->createMock('\phpbb\db\migrator');
 	}
 
 	/**
@@ -79,8 +74,7 @@ class ext_test extends \phpbb_test_case
 			->with('config')
 			->willReturn($config);
 
-		/** @var \vse\topicpreview\ext */
-		$ext = new \vse\topicpreview\ext($this->container, $this->extension_finder, $this->migrator, 'vse/topicpreview', '');
+		$ext = new ext($this->container, $this->extension_finder, $this->migrator, 'vse/topicpreview', '');
 
 		self::assertSame($expected, $ext->is_enableable());
 	}